+++ pointillism/README.md
@@ -1,0 +1,10 @@
+# Pointillism
+Image manipulation for various "pointillism" effects, built from scratch in Python. Image handling courtesy of Pillow. 
+
+Under active development. Please let me know if you have feedback.
+
+## Examples
+See usage and examples in [demo notebook](https://github.com/tsennott/pointillism/blob/master/Pointillism%20Example%20Usage.ipynb)
+
+## Deployment
+Currently deployed via Django on AWS Elastic beanstalk at [pointillizer.com](http://www.pointillizer.com)
