+++ pointillism/pointillism/__init__.py
@@ -1,0 +1,11 @@
+#! /usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Pointillism module
+"""
+
+from .main import main
+from .gif import gif
+from .batch import batch
+
+name = "pointillism"
